+++ lib/widgets/app_bar.dart
@@ -13,7 +13,7 @@ class _MyAppBarState extends State<MyAppBar> {
   @override
   Widget build(BuildContext context) {
     return AppBar(
-      backgroundColor: Colors.white,
+      backgroundColor: Colors.grey,
       title: (Text(
         widget.pageName,
         textAlign: TextAlign.right,
